+++ mobile/android/android-components/components/browser/storage-sync/src/main/java/mozilla/components/browser/storage/sync/Types.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-@file:Suppress("MatchingDeclarationName", "TooManyFunctions")
+@file:Suppress("MatchingDeclarationName")
 
 package mozilla.components.browser.storage.sync
 
@@ -11,20 +11,15 @@ import mozilla.appservices.places.BookmarkItem
 import mozilla.appservices.places.BookmarkSeparator
 import mozilla.appservices.places.BookmarkTreeNode
 import mozilla.appservices.places.SyncAuthInfo
-import mozilla.components.concept.storage.Address
 import mozilla.components.concept.storage.BookmarkNode
 import mozilla.components.concept.storage.BookmarkNodeType
-import mozilla.components.concept.storage.CreditCard
 import mozilla.components.concept.storage.FrecencyThresholdOption
 import mozilla.components.concept.storage.TopFrecentSiteInfo
 import mozilla.components.concept.storage.VisitInfo
 import mozilla.components.concept.storage.VisitType
-import mozilla.appservices.autofill.UpdatableAddressFields
-import mozilla.appservices.autofill.UpdatableCreditCardFields
 
-// We have type definitions at the concept level, and "external" types defined within
-// Autofill and Places. In practice these two types are largely the same, and this file
-// is the conversion point.
+// We have type definitions at the concept level, and "external" types defined within Places.
+// In practice these two types are largely the same, and this file is the conversion point.
 
 /**
  * Conversion from our SyncAuthInfo into its "native" version used at the interface boundary.
@@ -105,80 +100,4 @@ internal fun BookmarkTreeNode.asBookmarkNode(): BookmarkNode {
             BookmarkNode(BookmarkNodeType.SEPARATOR, this.guid, this.parentGUID, this.position, null, null, null)
         }
     }
-}
-
-/**
- * Conversion from a generic [UpdatableAddressFields] into its richer comrade within the 'autofill' lib.
- */
-internal fun mozilla.components.concept.storage.UpdatableAddressFields.into(): UpdatableAddressFields {
-    return UpdatableAddressFields(
-        givenName = this.givenName,
-        additionalName = this.additionalName,
-        familyName = this.familyName,
-        organization = this.organization,
-        streetAddress = this.streetAddress,
-        addressLevel3 = this.addressLevel3,
-        addressLevel2 = this.addressLevel2,
-        addressLevel1 = this.addressLevel1,
-        postalCode = this.postalCode,
-        country = this.country,
-        tel = this.tel,
-        email = this.email
-    )
-}
-
-/**
- * Conversion from a generic [UpdatableCreditCardFields] into its richer comrade within the 'autofill' lib.
- */
-internal fun mozilla.components.concept.storage.UpdatableCreditCardFields.into(): UpdatableCreditCardFields {
-    return UpdatableCreditCardFields(
-        ccName = this.billingName,
-        ccNumber = this.cardNumber,
-        ccExpMonth = this.expiryMonth,
-        ccExpYear = this.expiryYear,
-        ccType = this.cardType
-    )
-}
-
-/**
- * Conversion from a "native" autofill [Address] into its generic comrade.
- */
-internal fun mozilla.appservices.autofill.Address.into(): Address {
-    return Address(
-        guid = this.guid,
-        givenName = this.givenName,
-        additionalName = this.additionalName,
-        familyName = this.familyName,
-        organization = this.organization,
-        streetAddress = this.streetAddress,
-        addressLevel3 = this.addressLevel3,
-        addressLevel2 = this.addressLevel2,
-        addressLevel1 = this.addressLevel1,
-        postalCode = this.postalCode,
-        country = this.country,
-        tel = this.tel,
-        email = this.email,
-        timeCreated = this.timeCreated,
-        timeLastUsed = this.timeLastUsed,
-        timeLastModified = this.timeLastModified,
-        timesUsed = this.timesUsed
-    )
-}
-
-/**
- * Conversion from a "native" autofill [CreditCard] into its generic comrade.
- */
-internal fun mozilla.appservices.autofill.CreditCard.into(): CreditCard {
-    return CreditCard(
-        guid = this.guid,
-        billingName = this.ccName,
-        cardNumber = this.ccNumber,
-        expiryMonth = this.ccExpMonth,
-        expiryYear = this.ccExpYear,
-        cardType = this.ccType,
-        timeCreated = this.timeCreated,
-        timeLastUsed = this.timeLastUsed,
-        timeLastModified = this.timeLastModified,
-        timesUsed = this.timesUsed
-    )
 }